refactor: 패키지 이동 및 불필요한 개행 제거
